src/requirements.txt
# WAF Proxy 依赖
httpx[http2]>=0.27.0
fastapi>=0.109.0
uvicorn>=0.27.0
playwright>=1.40.0
loguru>=0.7.2
python-dotenv>=1.0.0
apscheduler>=3.10.0
pydantic>=2.0.0
